requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
install==1.3.4
joblib==0.17.0
kaleido==0.0.3.post1
nltk==3.5
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.4
plotly==4.12.0
pymystem3==0.1.10
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.4
pyzbar==0.1.8
regex==2020.9.27
requests==2.7.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.50.2
urllib3==1.25.10
